more tests for sum and tanh autograd (#261)
* more tests for sum and tanh autograd

* fix tests
