--- UnityProject/Assets/OpenMined.Tests/Editor/Autograd.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 401b9b031057e40e6a92e30f8560f493
-folderAsset: yes
-timeCreated: 1513805710
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
