--- UnityProject/Assets/OpenMined.Tests/Editor/Autograd/AutogradTest.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: e37eb85c7300440d88614f59804facd0
-timeCreated: 1513805711
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
